feat(signup) configure signup api for database >- change data structure to database [Delivers #166720121]
