+++ .babelrc
@@ -2,4 +2,7 @@
     "presets": [
         "@babel/preset-env"
     ],
+    "plugins": [
+        "@babel/plugin-transform-runtime"
+    ]
 }
