+++ package.json
@@ -8,8 +8,8 @@
     "createtest": "cross-env NODE_ENV=test babel-node server/db/migrations/createTables.js",
     "drop": " cross-env NODE_ENV=production babel-node server/db/migrations/dropTables.js",
     "droptest": "cross-env NODE_ENV=test babel-node server/db/migrations/dropTables.js",
-    "start": "rimraf dist/ && babel ./server -d dist && node dist/app.js",
-    "startdev": "nodemon --exec babel-node server/app.js",
+    "start": "cross-env NODE_ENV=production rimraf dist/ && babel ./server -d dist && node dist/app.js",
+    "startdev": "cross-env NODE_ENV=production nodemon --exec babel-node server/app.js",
     "test": "cross-env NODE_ENV=test nyc --require @babel/register mocha server/test --exit",
     "testdev": "npm run droptest && npm run createtest && cross-env NODE_ENV=test nyc --require @babel/register  mocha server/test --exit && npm run droptest",
     "coverage": "nyc report --reporter=text-lcov | coveralls",
@@ -30,6 +30,7 @@
   },
   "homepage": "https://github.com/amoskeyz/AutoMart#readme",
   "dependencies": {
+    "@babel/runtime": "^7.4.5",
     "bcrypt": "^3.0.6",
     "bcrypt-nodejs": "0.0.3",
     "body-parser": "^1.19.0",
@@ -50,6 +51,8 @@
     "@babel/cli": "^7.4.4",
     "@babel/core": "^7.4.5",
     "@babel/node": "^7.4.5",
+    "@babel/plugin-transform-regenerator": "^7.4.0",
+    "@babel/plugin-transform-runtime": "^7.4.0",
     "@babel/preset-env": "^7.4.5",
     "coveralls": "^3.0.3",
     "cross-env": "^5.2.0",
