+++ server.1/helper/encrypt.js
@@ -4,6 +4,10 @@ class Encrypt {
   static passwordhash(password) {
     return bcrypt.hashSync(password, 10);
   }
+
+  static compare(password, hash) {
+    return bcrypt.compareSync(password, hash);
+  }
 }
 
 export default Encrypt;
