+++ server.1/helper/token.js
@@ -1,0 +1,10 @@
+import jwt from 'jsonwebtoken';
+import dotenv from 'dotenv';
+
+dotenv.config();
+
+const token = payload => jwt.sign(payload, process.env.secretKey, {
+  expiresIn: '1d',
+});
+
+export default token;
