+++ server.1/router/userRoute.js
@@ -5,5 +5,6 @@ import validator from '../middleware/validator';
 const router = express();
 router.get('/', userController.welcome);
 router.post('/auth/signup', validator.validateSignup, userController.signupUser);
+router.post('/auth/signin', validator.validateSignin, userController.signinUser);
 
 export default router;
