+++ server.1/test/data/user.js
@@ -1,0 +1,24 @@
+const user = [
+  {
+    firstName: 'support',
+    lastName: 'automart',
+    email: 'support11@automart.com',
+    password: 'automart',
+    phoneNumber: '09087897656',
+  },
+  {
+    lastName: 'automart',
+    email: 'support@automart.com',
+    password: 'automart',
+    phoneNumber: '09087897656',
+  },
+  {
+    email: 'support@automart.com',
+    password: 'auto0mart',
+  },
+  {
+    email: 'admin@automart.com',
+    password: '12345678automart2019',
+  },
+];
+export default user;
